fix wavefield update

+++ src/utils/include/dataType.hpp
@@ -221,7 +221,7 @@ typedef Kokkos::View<int ***, Layout, MemSpace> array3DInt;
 typedef Kokkos::View<float ***, Layout, MemSpace> array3DReal;
 typedef Kokkos::View<double ***, Layout, MemSpace> array3DDouble;
 
-typedef Kokkos::View<int ****, Layout, MemSpace> array4DInt;
+typedef Kokkos::View<int64_t ****, Layout, MemSpace> array4DInt;
 typedef Kokkos::View<float ****, Layout, MemSpace> array4DReal;
 typedef Kokkos::View<double ****, Layout, MemSpace> array4DDouble;
